Moved chat handler threads to gamestate. N.B. The intention is to have multiple gamestates so this is not final. Changed wndow to resizable. This works but is visually untidy.
